Create Add_Subtract_Variable_Calculator
Using the Stack_8 emulator and code, write the following program:
Create a 4 function integer calculator that will clear variable m and can add, subtract to that variable, like so:
	Enter c, +, -, q: c
	M is: 0
Enter c, +, -, q: +
Number? 124
M is: 124
Enter c, +, -, q: -
Number? 3
M is 121
Enter c, +, -, q: q
	Done!

Use the I/O routines to have the input and output of decimal and print the message. Suggested to have a subroutine to do the prompt and input the number.

Extra credit, 10 pts
Add the multiply and divide (*, /).
You’ll have to use the MQ register.